Refactor: Fix process step icon truncation and update component names
