+++ src/components/HowItWorks-Metric.svelte
