+++ src/routes/+page.svelte
@@ -138,21 +138,21 @@
           abroad.'}
         tier={'Tier 1 (beginner)'}
         techStack={'HTML, CSS, Flexbox'}
-        images={[{ type: 'avif', img: '/landing/JobbaticalProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/JobbaticalProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/JobbaticalProject.png', alt: 'Voyage project screenshot' }]} />
+        images={[{ type: 'avif', img: '/landing/tier-projects/JobbaticalProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/tier-projects/JobbaticalProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/tier-projects/JobbaticalProject.png', alt: 'Voyage project screenshot' }]} />
       <LandingProjectExample
         name={'Chrome Extension App'}
         description={'A finance app to add income & expense categories \
           with graph rendering.'}
         tier={'Tier 2 (intermediate)'}
         techStack={'Frontend Framework & API usage'}
-        images={[{ type: 'avif', img: '/landing/BudgetBoardProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/BudgetBoardProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/BudgetBoardProject.png', alt: 'Voyage project screenshot' }]} />
+        images={[{ type: 'avif', img: '/landing/tier-projects/BudgetBoardProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/tier-projects/BudgetBoardProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/tier-projects/BudgetBoardProject.png', alt: 'Voyage project screenshot' }]} />
       <LandingProjectExample
         name={'Volunteer Manager App'}
         description={'Helps people to recruit and manage volunteers for a \
           non-profit project.'}
         tier={'Tier 3 (advanced)'}
-        techStack={'Backend API, any Database (optional)'}
-        images={[{ type: 'avif', img: '/landing/VolunteerAppProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/VolunteerAppProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/VolunteerAppProject.png', alt: 'Voyage project screenshot' }]} />
+        techStack={'Backend API, Database (optional)'}
+        images={[{ type: 'avif', img: '/landing/tier-projects/VolunteerAppProject.avif', alt: 'Voyage project screenshot' }, { type: 'webp', img: '/landing/tier-projects/VolunteerAppProject.webp', alt: 'Voyage project screenshot' }, { type: 'png', img: '/landing/tier-projects/VolunteerAppProject.png', alt: 'Voyage project screenshot' }]} />
     </div>
   </div>
 </div>
